# server/src/fastify-server.ts
import Fastify from "fastify";
import fastifySwagger from "@fastify/swagger";
import fastifySwaggerUi from "@fastify/swagger-ui";
import { Type } from "@sinclair/typebox";
import { Value } from "@sinclair/typebox/value";
import type { TypeBoxTypeProvider } from "@fastify/type-provider-typebox";

process.loadEnvFile(new URL("../.env", import.meta.url));

const PORT = 3000;

const SPOTIFY_FETCH_TIMEOUT_MS = 5000;

const SpotifyTokenResponse = Type.Object({
  access_token: Type.String(),
});

async function getSpotifyAccessToken(): Promise<string> {
  const clientId = process.env.SPOTIFY_CLIENT_ID;
  const clientSecret = process.env.SPOTIFY_CLIENT_SECRET;

  const credentials = Buffer.from(`${clientId}:${clientSecret}`).toString(
    "base64",
  );

  const response = await fetch("https://accounts.spotify.com/api/token", {
    method: "POST",
    headers: {
      "Content-Type": "application/x-www-form-urlencoded",
      Authorization: `Basic ${credentials}`,
    },
    body: "grant_type=client_credentials",
    signal: AbortSignal.timeout(SPOTIFY_FETCH_TIMEOUT_MS),
  });
  const rawData = await response.json();
  const data = Value.Parse(SpotifyTokenResponse, rawData);
  return data.access_token;
}

const HealthResponse = Type.Object({
  status: Type.String(),
});

const Track = Type.Object({
  id: Type.String(),
  name: Type.String(),
  artists: Type.Array(Type.String()),
  albumImageUrl: Type.Optional(Type.String()),
  previewUrl: Type.Optional(Type.String()),
});

const TrackListResponse = Type.Object({
  tracks: Type.Array(Track),
});

const TracksQuery = Type.Object({
  genre: Type.String({ minLength: 1 }),
  limit: Type.Optional(Type.Integer({ minimum: 1, maximum: 10, default: 10 }))
});

const BadRequest = Type.Object({
  statusCode: Type.Integer(),
  code: Type.String(),
  error: Type.String(),
  message: Type.String(),
});

const ErrorResponse = Type.Object({
  error: Type.String(),
});

const SpotifyTrackSearchItem = Type.Object({
  id: Type.String(),
  name: Type.String(),
  artists: Type.Array(Type.Object({ name: Type.String() })),
  album: Type.Object({
    images: Type.Array(Type.Object({ url: Type.String() })),
  }),
  preview_url: Type.Optional(Type.Union([Type.String(), Type.Null()])),
});

const SpotifyTrackSearchApiResponse = Type.Object({
  tracks: Type.Object({
    items: Type.Array(SpotifyTrackSearchItem),
  }),
});

async function searchSpotifyTracks(query: string, limit?: number) {
  const accessToken = await getSpotifyAccessToken();

  const params = new URLSearchParams({ q: query, type: "track" });
  if (limit !== undefined) {
    params.set("limit", String(limit));
  }

  const spotifyResponse = await fetch(
    `https://api.spotify.com/v1/search?${params}`,
    {
      headers: {
        Authorization: `Bearer ${accessToken}`,
      },
      signal: AbortSignal.timeout(SPOTIFY_FETCH_TIMEOUT_MS),
    },
  );
  const rawData = await spotifyResponse.json();

  const data = Value.Parse(SpotifyTrackSearchApiResponse, rawData);

  return data.tracks.items.map((item) => ({
    id: item.id,
    name: item.name,
    artists: item.artists.map((a) => a.name),
    albumImageUrl: item.album.images[0]?.url,
    previewUrl: item.preview_url ?? undefined,
  }));
}

export async function buildApp() {
  const fastify = Fastify({
    logger: true,
  }).withTypeProvider<TypeBoxTypeProvider>();

  await fastify.register(fastifySwagger, {
    openapi: {
      info: {
        title: "MoodHub Server API",
        version: "0.1.0",
      },
    },
  });

  await fastify.register(fastifySwaggerUi, {
    routePrefix: "/documentation",
  });

  fastify.get(
    "/health",
    {
      schema: {
        response: {
          200: HealthResponse,
        },
      },
    },
    (request, reply) => {
      return { status: "ok" };
    },
  );

  fastify.get(
    "/search",
    {
      schema: {
        querystring: Type.Object({
          artist: Type.String(),
        }),
        response: {
          200: TrackListResponse,
          400: BadRequest,
          502: ErrorResponse,
        },
      },
    },
    async (request, reply) => {
      const { artist } = request.query;
      try {
        const tracks = await searchSpotifyTracks(`artist:${artist}`);
        return { tracks };
      } catch (error) {
        reply
          .code(502)
          .send({ error: "Spotifyから予期しない形式のレスポンスが返されました" });
      }
    },
  );

  fastify.get(
    "/tracks",
    {
      schema: {
        querystring: TracksQuery,
        response: {
          200: TrackListResponse,
          400: BadRequest,
          502: ErrorResponse,
        },
      },
    },
    async (request, reply) => {
      const { genre, limit } = request.query;
      try {
        const tracks = await searchSpotifyTracks(`genre:${genre}`, limit);
        return { tracks };
      } catch (error) {
        reply
          .code(502)
          .send({ error: "Spotifyから予期しない形式のレスポンスが返されました" });
      }
    },
  );

  return fastify;
}

if (import.meta.main) {
  if (!process.env.SPOTIFY_CLIENT_ID || !process.env.SPOTIFY_CLIENT_SECRET) {
    throw new Error(
      "環境変数 SPOTIFY_CLIENT_ID / SPOTIFY_CLIENT_SECRET が設定されていません",
    );
  }

  const fastify = await buildApp();
  fastify.listen({ port: PORT }, (err) => {
    if (err) {
      fastify.log.error(err);
      process.exit(1);
    }
  });
}